% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/currency_tools.R
\docType{package}
\name{currency_tools}
\alias{janitorID}
\alias{janitorID-package}
\alias{currency_tools}
\title{Currency Conversion Tools with Real-Time Exchange Rates}
\description{
Mendukung 146 mata uang dari 5 benua: Asia (48), Eropa (44), Afrika (54),
Amerika (35), dan Oseania (14).
}
\author{
\strong{Maintainer}: Kelompok Data Science 1 \email{ds1@example.com} (\href{https://orcid.org/0000-0000-0000-0000}{ORCID})

Authors:
\itemize{
  \item Kelompok Data Science 1 \email{ds1@example.com} (\href{https://orcid.org/0000-0000-0000-0000}{ORCID})
}

}
\keyword{internal}
