% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_analyze.R
\name{plot_categorical_presentation}
\alias{plot_categorical_presentation}
\title{Plot Distribusi Kategorik untuk Presentasi}
\usage{
plot_categorical_presentation(data)
}
\arguments{
\item{data}{Data frame yang akan dianalisis}
}
\value{
Objek ggplot
}
\description{
Plot Distribusi Kategorik untuk Presentasi
}
