% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data_health.R
\name{plot_data_health}
\alias{plot_data_health}
\title{Dashboard Kesehatan Data Interaktif Lengkap}
\usage{
plot_data_health(
  data,
  interactive = FALSE,
  show_outliers = TRUE,
  show_correlation = TRUE,
  show_missing = TRUE,
  show_distribution = TRUE,
  show_summary = TRUE,
  title = NULL,
  max_cols = 15,
  palette = "default",
  is_cleaned = FALSE
)
}
\arguments{
\item{data}{Data frame yang akan dianalisis}

\item{interactive}{Boolean, jika TRUE menggunakan plotly (interaktif)}

\item{show_outliers}{Boolean, menampilkan plot outlier}

\item{show_correlation}{Boolean, menampilkan heatmap korelasi}

\item{show_missing}{Boolean, menampilkan pola data hilang}

\item{show_distribution}{Boolean, menampilkan grafik distribusi}

\item{show_summary}{Boolean, menampilkan summary cards}

\item{title}{Karakter, judul utama dashboard}

\item{max_cols}{Integer, batas maksimal kolom yang diproses}

\item{palette}{Karakter, pilihan warna ("default", "colorblind", "vibrant")}

\item{is_cleaned}{Boolean, jika TRUE grafik outlier dipaksa 0\%}
}
\value{
Objek ggplot atau plotly
}
\description{
Fungsi ini menghasilkan dashboard komprehensif untuk mengecek kesehatan data:
\itemize{
\item Pola missing values
\item Deteksi outlier
\item Heatmap korelasi
\item Distribusi numerik & kategorik
\item Ringkasan data
}
}
