% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/currency_tools.R
\docType{package}
\name{currency_tools}
\alias{janitorID}
\alias{janitorID-package}
\alias{currency_tools}
\title{Currency Conversion Tools with Real-Time Exchange Rates}
\description{
Mendukung 146 mata uang dari 5 benua: Asia (48), Eropa (44), Afrika (54),
Amerika (35), dan Oseania (14).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/nnnaaakar209-sketch/janitorID}
  \item Report bugs at \url{https://github.com/nnnaaakar209-sketch/janitorID/issues}
}

}
\author{
\strong{Maintainer}: Kelompok Data Science 1 \email{janitorid.team@example.com} (\href{https://orcid.org/0009-0000-0000-0000}{ORCID})

Authors:
\itemize{
  \item Kelompok Data Science 1 \email{janitorid.team@example.com} (\href{https://orcid.org/0009-0000-0000-0000}{ORCID})
}

Other contributors:
\itemize{
  \item Sam Firke (Original janitor package author) [contributor]
}

}
\keyword{internal}
